Fixed for Windows.
Need to construct the path for Windows rather than using env variables
the way were we doing it before.

Move all functions that need path and config info to a single file,
config.js.

Detached should be false, esp for windows.

Don't use fd 4 that doesn't seem to work on win with wallet.

Closes #122
