+++ app/config.js
@@ -1,3 +1,7 @@
+import fs from 'fs';
+import path from 'path';
+import os from 'os';
+
 export function getCfg() {
   const Config = require('electron-config');
   const config = new Config();
@@ -48,4 +52,67 @@ export function getCfgPath() {
   const Config = require('electron-config');
   const config = new Config();
   return config.path;
+}
+
+export function appDataDirectory() {
+  const path = require('path');
+  const os = require('os');
+
+  if (os.platform() == 'win32') {
+    return path.join(os.homedir(), 'AppData', 'Local', 'Decrediton');
+  } else if (process.platform === 'darwin') {
+    return path.join(os.homedir(), 'Library','Application Support','decrediton');
+  } else {
+    return path.join(os.homedir(),'.config','decrediton');
+  }
+}
+
+export function getCert() {
+  var cert = '';
+  var cfg = getCfg();
+  if (cfg.cert_path != '') {
+    return(cfg.cert_path);
+  }
+  var certPath = '';
+  if (os.platform() == 'win32') {
+    certPath = path.join(os.homedir(), 'AppData', 'Local', 'Decrediton', 'rpc.cert');
+  } else if (os.platform() == 'darwin') {
+    certPath = path.join(process.env.HOME, 'Library', 'Application Support',
+            'decrediton', 'rpc.cert');
+  } else {
+    certPath = path.join(process.env.HOME, '.config', 'decrediton', 'rpc.cert');
+  }
+
+  try {
+    cert = fs.readFileSync(certPath);
+  } catch (err) {
+    if (err.code === 'ENOENT') {
+      console.log(certPath + ' does not exist');
+    } else if (err.code === 'EACCES') {
+      console.log(certPath + ' permission denied');
+    } else {
+      console.error(certPath + ' ' + err);
+    }
+  }
+
+  return(cert);
+}
+
+export function getDcrdCert() {
+  var cfg = getCfg();
+  if (cfg.daemon_cert_path != '') {
+    return(cfg.daemon_cert_path);
+  }
+  var certPath = '';
+  if (os.platform() == 'win32') {
+    certPath = path.join(os.homedir(), 'AppData', 'Local', 'Dcrd', 'rpc.cert');
+  } else if (os.platform() == 'darwin') {
+    certPath = path.join(process.env.HOME, 'Library', 'Application Support',
+            'Dcrd', 'rpc.cert');
+  } else {
+    certPath = path.join(process.env.HOME, '.dcrd', 'rpc.cert');
+  }
+
+  var cert = fs.readFileSync(certPath);
+  return(cert);
 }
